requirements.txt
dnspython>=2.1.0
eth-keys>=0.3.3
parse>=1.19.0
rlp>=2.0.1
eth-hash[pycryptodome]>=0.3.1
trio>=0.18.0
trio-typing>=0.5.1
lru-dict>=1.1.7
eth-utils>=1.10.0
python-snappy>=0.6.0
cryptography>=3.1.1
ujson>=4.0.1
requests>=2.26.0
python-logstash>=0.4.6
hypercorn>=0.11.2
quart-trio>=0.8.0
httpx>=0.18.2